# packages/service/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactList": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "title": "Artifacts",
            "type": "array"
          }
        },
        "required": [
          "artifacts"
        ],
        "title": "ArtifactList",
        "type": "object"
      },
      "ArtifactResponse": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "kind": {
            "enum": [
              "ttf",
              "woff2",
              "proof_html",
              "qa_json"
            ],
            "title": "Kind",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "job_id",
          "bytes",
          "sha256",
          "created_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "GlyphList": {
        "properties": {
          "glyphs": {
            "items": {
              "$ref": "#/components/schemas/GlyphResponse"
            },
            "title": "Glyphs",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "glyphs"
        ],
        "title": "GlyphList",
        "type": "object"
      },
      "GlyphResponse": {
        "properties": {
          "advance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advance"
          },
          "char": {
            "title": "Char",
            "type": "string"
          },
          "codepoint": {
            "title": "Codepoint",
            "type": "string"
          },
          "status": {
            "enum": [
              "missing",
              "auto",
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "svg_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Url"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "enum": [
                "LOW_INK",
                "TOUCHES_BORDER",
                "TINY_CONTOURS_REMOVED",
                "LARGE_INK_BLOB",
                "OFF_GUIDE"
              ],
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "codepoint",
          "char",
          "status",
          "warnings",
          "updated_at"
        ],
        "title": "GlyphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewError": {
        "properties": {
          "codepoint": {
            "title": "Codepoint",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "codepoint",
          "reason"
        ],
        "title": "ReviewError",
        "type": "object"
      },
      "ReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "accept": {
            "items": {
              "type": "string"
            },
            "maxItems": 400,
            "title": "Accept",
            "type": "array"
          },
          "reject": {
            "items": {
              "type": "string"
            },
            "maxItems": 400,
            "title": "Reject",
            "type": "array"
          }
        },
        "title": "ReviewRequest",
        "type": "object"
      },
      "ReviewResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ReviewError"
            },
            "title": "Errors",
            "type": "array"
          },
          "unchanged": {
            "title": "Unchanged",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "updated",
          "unchanged",
          "errors"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "glyphlab API",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/artifacts": {
      "get": {
        "operationId": "list_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Artifacts"
      }
    },
    "/api/projects/{project_id}/artifacts/{artifact_id}": {
      "get": {
        "operationId": "download_artifact",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "findings": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "check_id": {
                            "type": "string"
                          },
                          "detail": {
                            "type": [
                              "object",
                              "null"
                            ]
                          },
                          "message": {
                            "type": "string"
                          },
                          "severity": {
                            "enum": [
                              "FAIL",
                              "WARN",
                              "SKIPPED",
                              "ALLOWED"
                            ]
                          }
                        },
                        "required": [
                          "check_id",
                          "severity",
                          "message",
                          "detail"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "passed": {
                      "type": "boolean"
                    },
                    "schema": {
                      "const": "glyphlab.qa-report/1"
                    }
                  },
                  "required": [
                    "schema",
                    "passed",
                    "findings"
                  ],
                  "type": "object"
                }
              },
              "font/ttf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "font/woff2": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "text/html": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download Artifact"
      }
    },
...
    "/api/projects/{project_id}/glyphs": {
      "get": {
        "operationId": "list_glyphs",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "missing",
                      "auto",
                      "accepted",
                      "rejected"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "warning",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "LOW_INK",
                    "TOUCHES_BORDER",
                    "TINY_CONTOURS_REMOVED",
                    "LARGE_INK_BLOB",
                    "OFF_GUIDE"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warning"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": -1,
              "maximum": 1114111,
              "minimum": -1,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 300,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlyphList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Glyphs"
      }
    },
    "/api/projects/{project_id}/glyphs/{cp}.svg": {
      "get": {
        "operationId": "get_glyph_svg",
        "parameters": [
          {
            "in": "path",
            "name": "cp",
            "required": true,
            "schema": {
              "title": "Cp",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/svg+xml": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Glyph Svg"
      }
    },
    "/api/projects/{project_id}/glyphs:review": {
      "post": {
        "operationId": "review_glyphs",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Review Glyphs"
      }
    },
...
  }
}